+++ examples/event-read.rs
@@ -5,7 +5,7 @@ use std::io::{stdout, Write};
 
 use crossterm::{
     cursor::position,
-    event::{read, DisableMouseCapture, EnableMouseCapture, Event, KeyEvent},
+    event::{read, DisableMouseCapture, EnableMouseCapture, Event, KeyCode},
     execute,
     screen::RawScreen,
     Result,
@@ -24,11 +24,11 @@ fn print_events() -> Result<()> {
 
         println!("Event::{:?}\r", event);
 
-        if event == Event::Key(KeyEvent::Char('c')) {
+        if event == Event::Key(KeyCode::Char('c').into()) {
             println!("Cursor position: {:?}\r", position());
         }
 
-        if event == Event::Key(KeyEvent::Esc) {
+        if event == Event::Key(KeyCode::Esc.into()) {
             break;
         }
     }